feat(origin-secrets): v0.4.2 — green pipeline gates, fix init passphrase routing
- Fix HIGH bug: init ignored the global -p/--passphrase-file flag and
  silently encrypted the vault with a demo string, causing every later
  command to fail with VaultDecryptionFailed. Passphrase now flows through
  a single global path shared by all subcommands (cmd_init signature
  updated; subcommand-local -p removed).
- Gate test helper modules (audit_tests/dispatch_tests/share_tests)
  behind #[cfg(test)] so test code no longer ships in the production lib.
- Clear clippy -D warnings on origin-secrets (rename MemoryTier::from_str
  to parse_tier; apply cargo clippy --fix lints).
- CI: clippy job now denies warnings on -p origin-secrets and runs a
  non-failing cargo clippy --workspace pass for visibility, so the SDK's
  intentional dead-code internals no longer block the pipeline.
- Bump crate version to 0.4.2 (own version, ahead of workspace 0.4.1).
- Add CHANGELOG.md.

+++ origin-secrets/src/share_tests.rs
@@ -1,7 +1,6 @@
 //! Unit tests for Share structures
 
 use crate::share::{HybridSignature, Share};
-use serde_json;
 
 #[test]
 fn test_share_creation() {
@@ -54,7 +53,10 @@ fn test_share_serialization_roundtrip() {
     assert_eq!(share.share_data, deserialized.share_data);
     assert_eq!(share.recipient, deserialized.recipient);
     assert_eq!(share.signature.ed25519, deserialized.signature.ed25519);
-    assert_eq!(share.signature.falcon1024, deserialized.signature.falcon1024);
+    assert_eq!(
+        share.signature.falcon1024,
+        deserialized.signature.falcon1024
+    );
 }
 
 #[test]
@@ -201,11 +203,11 @@ fn test_share_various_numbers() {
             share_number: i,
             threshold: 3,
             total_shares: 5,
-            share_data: vec![i as u8; 10],
+            share_data: vec![i; 10],
             fingerprint: format!("fp-{}", i),
             signature: HybridSignature {
-                ed25519: vec![i as u8; 5],
-                falcon1024: vec![(i + 1) as u8; 5],
+                ed25519: vec![i; 5],
+                falcon1024: vec![i + 1; 5],
             },
             created_at: "2026-07-30T21:27:45Z".to_string(),
             recipient: Some(format!("share-{}.example.com", i)),
